feat: added the global sensor file that adds the Sensor class
This class is the mother class of all sensors and it will be the one
stored with virtual functions to call different child functions.

For example if you create a RawSensor, it will be stored as a Sensor
and accessing it's raw_read method will be through the virtual
raw_read method that will be overwritten by the RawSensor class.
If a function is irrelevant in the mother class, or other classes, it
will be defaulted by an error message.